apply request from reviewer
